Warn when converting non-utf-8 data
Marc data is usually encoded as marc8 or utf-8. Vim does not understand
marc8, so it misinterprets it as a different encoding. This causes data
loss when we try to convert it.

I'm not sure if we can make Vim understand marc8, or if that would even
be worth the effort. So we'll just warn the user if they inadvertently
use it.

> "MARC8 is essentially a made up encoding — it simply doesn’t exist
  outside of the small world of MARC21 libraries."
  -- Terry Reese, <https://blog.reeset.net/archives/1897>